requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.8.0
aiohttp==3.9.5
pytest==8.3.2
httpx==0.27.0
pytest-asyncio==0.23.8
firecrawl-py==4.4.0
pymupdf==1.23.8
python-multipart==0.0.6
langchain-text-splitters==0.3.0
openai>=1.12.0
